✨ feat: 인증처리 및 예외처리 기능 개선
- `ErrorMessage` enum에 필드 추가
- `UnauthorizedException` 클래스: 로그인이 되지 않은 상태에서 보호된 기능 호출 시 예외 처리하는 클래스
- `UserController` 클래스:
  - `UnauthorizedException` 예외처리 기능 추가
  - 탈퇴 시 세션 무효화되도록 수정
- `ScheduleController` 클래스:
  - 로그인했을 때만 작동하도록 인증 처리 기능 추가
  - `UnauthorizedException` 예외처리 기능 추가
- `UserService` 클래스:
  - 유저 정보 수정 시 이미 존재하는 이메일인 경우 예외 처리하도록 수정
- `ScheduleService` 클래스:
  - 일정 조회/수정/삭제 시 해당 일정이 유저가 생성한 일정인지 확인하는 로직 추가
- `ScheduleRepository` 클래스:
  - 일정 ID와 유저가 모두 일치하는 일정을 조회하는 커스텀 쿼리 메서드 추가
  - 일정 ID와 유저가 모두 일치하는 일정 존재를 확인하는 커스텀 쿼리 메서드 추가

+++ src/main/java/project/user/controller/UserController.java
@@ -2,6 +2,7 @@ package project.user.controller;
 
 import jakarta.servlet.http.HttpSession;
 import jakarta.validation.Valid;
+import project.config.UnauthorizedException;
 import project.user.dto.*;
 import project.user.service.UserService;
 import lombok.RequiredArgsConstructor;
@@ -33,43 +34,47 @@ public class UserController {
      * 로그인
      * @param request 로그인 Request DTO
      * @param session 유저의 로그인 상태
-     * @return 200(OK) 상태 코드
+     * @return 세션 정보와 200(OK) 상태 코드
      */
     @PostMapping("/login")
-    public ResponseEntity<String> login(
+    public ResponseEntity<SessionUser> login(
             @Valid @RequestBody LoginRequest request,
             HttpSession session
     ) {
         SessionUser sessionUser = userService.login(request); // 로그인한 유저 정보를 담는 DTO
         session.setAttribute("loginUser", sessionUser); // key: "loginUser", value: sessionUser
 
-        return ResponseEntity.status(HttpStatus.OK).body("로그인이 완료되었습니다.");
+        return ResponseEntity.status(HttpStatus.OK).body(sessionUser);
     }
 
     /**
      * 로그아웃
      * @param sessionUser 로그인한 유저 정보를 담은 DTO
      * @param session 유저의 로그인 상태
-     * @return 204(NO_CONTENT) 상태 코드, 로그인 안 했을 경우 400(Bad Request) 상태 코드
+     * @return 성공 시 204(NO_CONTENT) 상태 코드, 로그인 안 했을 경우 401(Unauthorized) 상태 코드
      */
     @PostMapping("/logout")
-    public ResponseEntity<String> logout(
-            @SessionAttribute(name = "loginUser") SessionUser sessionUser,
+    public ResponseEntity<Void> logout(
+            @SessionAttribute(name = "loginUser", required = false) SessionUser sessionUser,
             HttpSession session
     ) {
+        if (sessionUser == null) throw new UnauthorizedException();
+
         session.invalidate(); // 세션을 무효화
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).body("로그아웃이 완료되었습니다.");
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 
     /**
      * 유저 선택 조회
      * @param sessionUser 로그인한 유저 정보를 담은 DTO
-     * @return 유저 Response DTO와 200(OK) 상태 코드
+     * @return 유저 Response DTO와 200(OK) 상태 코드, 로그인 안 했을 경우 401(Unauthorized) 상태 코드
      */
     @GetMapping("/users")
     public ResponseEntity<UserResponse> getOneUser(
-            @SessionAttribute(name = "loginUser") SessionUser sessionUser
+            @SessionAttribute(name = "loginUser", required = false) SessionUser sessionUser
     ) {
+        if (sessionUser == null) throw new UnauthorizedException();
+
         UserResponse result = userService.getOne(sessionUser.getId());
         return ResponseEntity.status(HttpStatus.OK).body(result);
     }
@@ -88,27 +93,33 @@ public class UserController {
      * 유저 수정
      * @param sessionUser 로그인한 유저 정보를 담은 DTO
      * @param request 유저 수정 Request DTO
-     * @return 수정된 유저 Response DTO와 200(OK) 상태 코드
+     * @return 수정된 유저 Response DTO와 200(OK) 상태 코드, 로그인 안 했을 경우 401(Unauthorized) 상태 코드
      */
     @PutMapping("/users")
     public ResponseEntity<UserResponse> updateUser(
-            @SessionAttribute(name = "loginUser") SessionUser sessionUser,
+            @SessionAttribute(name = "loginUser", required = false) SessionUser sessionUser,
             @Valid @RequestBody UpdateUserRequest request
     ) {
+        if (sessionUser == null) throw new UnauthorizedException();
+
         UserResponse result = userService.update(sessionUser.getId(), request);
         return ResponseEntity.status(HttpStatus.OK).body(result);
     }
 
     /**
-     * 유저 삭제
+     * 유저 삭제: 탈퇴
      * @param sessionUser 로그인한 유저 정보를 담은 DTO
-     * @return 204(NO_CONTENT) 상태 코드
+     * @return 204(NO_CONTENT) 상태 코드, 로그인 안 했을 경우 401(Unauthorized) 상태 코드
      */
     @DeleteMapping("/users")
-    public ResponseEntity<String> deleteUser(
-            @SessionAttribute(name = "loginUser") SessionUser sessionUser
+    public ResponseEntity<Void> deleteUser(
+            @SessionAttribute(name = "loginUser", required = false) SessionUser sessionUser,
+            HttpSession session
     ) {
+        if (sessionUser == null) throw new UnauthorizedException();
+
         userService.delete(sessionUser.getId());
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).body("삭제가 완료되었습니다.");
+        session.invalidate(); // 세션을 무효화
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
     }
 }
